raising the state for track playing on all pages
