+++ src/components/basicPage/barContentForm.jsx
@@ -1,12 +1,14 @@
 import BarPlayer from './barPlayer'
 import * as S from '../styles/basicPage/basicPageStyles'
 
-const BarContentForm = ({ setIsPlaying, isPlaying }) => {
+const BarContentForm = ({ setIsPlaying, isPlaying, currentTrack, audioRef }) => {
   return (
     <S.Bar>
       <BarPlayer
+        audioRef={audioRef}
         isPlaying={isPlaying}
         setIsPlaying={setIsPlaying}
+        currentTrack={currentTrack}
         />
     </S.Bar>
   )
