+++ src/pages/appRender.jsx
@@ -1,28 +1,13 @@
-import { useState } from "react";
-import BarContentForm from "../components/basicPage/barContentForm";
-import { Container } from "../components/mainWrappers/container";
 import Main from "../components/mainWrappers/main";
-import { selectCurrentTrack } from "../store/selectors/tracks";
-import { useSelector } from "react-redux";
 
-function AppRender({onLogout}) {
-// передали трек в стор, сначала трек был в стейтах
-  const currentTrack = useSelector(selectCurrentTrack);
-  const [isPlaying, setIsPlaying] = useState(true);
+function AppRender({isPlaying, currentTrack, setIsPlaying}) {
 
+  
   return (
-    <Container>
       <Main
-        onLogout={onLogout}
+        currentTrack={currentTrack}
         isPlaying={isPlaying}
         setIsPlaying={setIsPlaying} />
-      {currentTrack &&
-        <BarContentForm
-        onLogout={onLogout}
-        isPlaying={isPlaying}
-        setIsPlaying={setIsPlaying} />
-       }
-    </Container>
   )
    
   }
