+++ src/routes.jsx
@@ -8,22 +8,23 @@ import { useAuthContext } from "./context/AuthContext"
 import { Layout } from "./pages/pageLayout/Layout"
 import FavoriteTracks from "./pages/favorites/favoriteTracks"
 
-
 export const AppRoutes = ({}) => {
     const { user } = useAuthContext()
+
+
 return (
     <Routes>
         <Route path="/login" element={<AuthPage />} />
         <Route path="/register" element={<AuthPage />} />
         <Route element={<ProtectedRoute isAllowed={Boolean(user)} />}>
-        <Route path="/" element={<Layout />}>
+        <Route path="/" element={<Layout/>}>
         <Route index element={
-        <Main/>}/>
+        <Main />}/>
         <Route path="/category/:id" element={<Playlist />}
         />
         <Route path="/favorites" element={<FavoriteTracks />} 
         />
-    </Route>
+        </Route>
     </Route>
             <Route path="*" element={<NotFound />} /> 
     </Routes>
