+++ src/store/selectors/tracks.js
@@ -8,4 +8,3 @@ export const selectCurrentTrackIndex = (state) => state.player.currentTrackIndex
 export const selectShuffledTracks = (state) => state.player.shuffledTracks
 export const selectIsLoop = (state) => state.player.isLoop
 export const selectFavoriteTracks = (state) => state.player.favoriteTracks
-console.log(selectFavoriteTracks);
